feat(packaging): restore AUR source package
