+++ rust/limux-ghostty-sys/Cargo.toml
@@ -5,6 +5,3 @@ edition.workspace = true
 authors.workspace = true
 license.workspace = true
 build = "build.rs"
-
-[build-dependencies]
-cc = "1"
