+++ rust/limux-host-linux/Cargo.toml
@@ -29,3 +29,6 @@ shell-quote = { version = "0.7.2", default-features = false, features = ["bash"]
 
 [dev-dependencies]
 tempfile.workspace = true
+
+[build-dependencies]
+cc = "1"
