+++ rust/limux-host-linux/src/main.rs
@@ -181,6 +181,8 @@ fn gtk_runtime_at_least(major: u32, minor: u32, micro: u32) -> bool {
 }
 
 fn main() {
+    limux_ghostty_sys::ensure_glad_symbols_linked();
+
     // Handle --version flag
     if std::env::args().any(|a| a == "--version" || a == "-v") {
         println!("Limux {VERSION}");
