+++ scripts/check.sh
@@ -20,3 +20,4 @@ cargo clippy --workspace --all-targets -- -D warnings
 cargo test --workspace -- --test-threads=1
 ./scripts/tests/test-release-version.sh
 ./scripts/tests/test-package-svg-loader.sh
+./scripts/tests/test-aur-source-package.sh
